refactor: split jpx-engine lib.rs into focused modules
Extract evaluation, introspection, and JSON utility code from lib.rs
(1,989 lines) into dedicated modules:

- eval.rs: evaluate, evaluate_str, batch_evaluate, validate
- introspection.rs: function search, describe, similar, categories
- json_utils.rs: format, diff, patch, merge, keys, paths, stats

lib.rs retains the JpxEngine struct definition, constructors, accessors,
query store delegation, and discovery delegation. Struct fields changed
to pub(crate) for cross-module impl blocks.

No public API changes. All 70 unit tests and 37 MCP integration tests
pass.

Closes #26
